Flashlight scripting
Worked on the flashlight script and got it to degrade down nicely.
Then worked on getting it to "recharge" when F is pressed which sort of works but it doesn't follow chargeSpeed like I want it to.
The flashlight degrading method also has a problem where using the current intensity breaks it.